+++ src/lib/zkapp.ts
@@ -1399,49 +1399,24 @@ Use the optional \`maxTransactionsWithActions\` argument to increase this number
     }): A[][] {
       let actionsForAccount: A[][] = [];
       Circuit.asProver(() => {
-        // if the fromActionHash is the empty state, we fetch all events
-        fromActionHash = fromActionHash
-          ?.equals(SequenceEvents.emptySequenceState())
-          .toBoolean()
-          ? undefined
-          : fromActionHash;
-
-        // used to determine start and end values in string
-        let start: string | undefined = fromActionHash
-          ? Ledger.fieldToBase58(fromActionHash)
-          : undefined;
-        let end: string | undefined = endActionHash
-          ? Ledger.fieldToBase58(endActionHash)
-          : undefined;
-
         let actions = Mina.getActions(
           contract.address,
           {
-            fromActionHash,
-            endActionHash,
+            fromActionState: fromActionHash,
+            endActionState: endActionHash,
           },
           contract.self.tokenId
         );
 
-        // gets the start/end indices of our array slice
-        let startIndex = start
-          ? actions.findIndex((e) => e.hash === start) + 1
-          : 0;
-        let endIndex = end
-          ? actions.findIndex((e) => e.hash === end) + 1
-          : undefined;
-
-        // slices the array so we only get the wanted range between fromActionHash and endActionHash
-        actionsForAccount = actions
-          .slice(startIndex, endIndex === 0 ? undefined : endIndex)
-          .map((event: { hash: string; actions: string[][] }) =>
-            // putting our string-Fields back into the original action type
+        // putting our string-Fields back into the original action type
+        actionsForAccount = actions.map(
+          (event: { hash: string; actions: string[][] }) =>
             event.actions.map((action: string[]) =>
               (reducer.actionType as ProvablePure<A>).fromFields(
                 action.map((fieldAsString: string) => Field(fieldAsString))
               )
             )
-          );
+        );
       });
 
       return actionsForAccount;